move shuffleArray to common dir
